fix: ingestion for XML streaming mode
There is a bug when you are trying to ingest `[{"Data": "Sample1"}, {"Data": "Sample2"}]`
it becomes `[{"Data": "Sample1"}, {"Data": "Sample2"}%5D` in Splunk because of the
`splunktalib`'s `escape_cdata` function used.

This PR removes usage of this function and uses `xml.sax.saxutils.escape` function to escape
all the characters before sending to Splunk.

P.S. You can play with it using this repository:
https://github.com/artemrys/splunk-xml-streaming-minimal-example.