BREAKING CHANGE: remove requests, splunktalib and data_collection module (#333)
`splunktaucclib` version 7.0.0 has just been released.

This BREAKING CHANGE release includes 3 major changes:

**1. `splunktaucclib` no longer provides `requests` library.** Instead,
it depends on the `requests` from the running Splunk environment - by
default it is the `requests` library shipped with Splunk. In case the
Splunk’s `requests` version is not sufficient for you,
you can deliver version you need by simply adding it to the
`requirements.txt` or `pyproject.toml` file in your add-on.

**2. `splunktaucclib` no longer depends on `splunktalib`.** Since
`splunktalib` is deprecated and archived, all relevant features have
been moved directly into `splunktaucclib` and `solnlib`. This means that
`splunktalib` is no longer required for add-ons with modular inputs.

**3. `splunktaucclib` no longer has `data_collection` module.** The
reason for this change is that this module should not be used in TA and
there is no migration path available. It is also worth mentioning that
`splunktaucclib` is now focused entirely on REST support and no other
functionalities will be available.

Ad. 1 and 2 - this change should be transparent for add-ons, especially
if you are using the latest version of Splunk. Despite the above it is
recommended to perform appropriate tests and possibly change the
libraries provided by the add-on.

Ad 3. If you use `data_collection` module in your TA - please reach out
to us in the user groups Slack channel.